Revert "Merge pull request #1102 from rocket-meals/codex/translate-and-style-done-button"
This reverts commit 2d3bc8b3af4bdda9aab4eb091e0d87ccd0b6b7cc, reversing
changes made to b3a2518fdfb0ddc6ade862d0a8a6260caae8c360.
