+++ apps/frontend/app/components/ProjectButton/index.tsx
@@ -1,30 +1,21 @@
-import React, { ReactNode } from 'react';
-import { Appearance, StyleProp, StyleSheet, Text, TouchableOpacity, ViewStyle } from 'react-native';
+import React from 'react';
+import { Appearance, StyleSheet, Text, TouchableOpacity } from 'react-native';
 import { useSelector } from 'react-redux';
 import { RootState } from '@/redux/reducer';
 import { myContrastColor } from '@/helper/ColorHelper';
 import { darkTheme, lightTheme } from '@/styles/themes';
+import { ProjectButtonProps } from './types';
 
-export interface ProjectButtonProps {
-	text: string;
-	onPress?: () => void;
-	iconLeft?: ReactNode;
-	iconRight?: ReactNode;
-	backgroundColor?: string;
-	style?: StyleProp<ViewStyle>;
-}
-
-const ProjectButton: React.FC<ProjectButtonProps> = ({ text, onPress, iconLeft, iconRight, backgroundColor, style }) => {
+const ProjectButton: React.FC<ProjectButtonProps> = ({ text, onPress, iconLeft, iconRight, style }) => {
 	const { primaryColor, selectedTheme } = useSelector((state: RootState) => state.settings);
 
 	const colorScheme = Appearance.getColorScheme();
 	const theme = selectedTheme === 'systematic' ? (colorScheme === 'dark' ? darkTheme : lightTheme) : selectedTheme === 'dark' ? darkTheme : lightTheme;
 
-	const resolvedBackgroundColor = backgroundColor ?? primaryColor;
-	const contrastColor = myContrastColor(resolvedBackgroundColor, theme, selectedTheme === 'dark');
+	const contrastColor = myContrastColor(primaryColor, theme, selectedTheme === 'dark');
 
 	return (
-		<TouchableOpacity style={[styles.container, { backgroundColor: resolvedBackgroundColor }, style]} onPress={onPress}>
+		<TouchableOpacity style={[styles.container, { backgroundColor: primaryColor }, style]} onPress={onPress}>
 			{iconLeft}
 			<Text style={[styles.label, { color: contrastColor }]}>{text}</Text>
 			{iconRight}
